+++ internal/lib/github_release_test.go
@@ -108,31 +108,6 @@ func TestGithubReleaseExecute(t *testing.T) {
 	})
 }
 
-func TestTagNormalization(t *testing.T) {
-	testData := []struct {
-		name string
-		tag  string
-		want string
-	}{
-		{
-			name: "kustomize",
-			tag:  "kustomize/v4.5.4",
-			want: "kustomize-v4.5.4",
-		},
-		{
-			name: "regular",
-			tag:  "v0.21.0",
-			want: "v0.21.0",
-		},
-	}
-	for _, tc := range testData {
-		t.Run(tc.name, func(t *testing.T) {
-			g := GithubRelease{Tag: tc.tag}
-			require.Equal(t, tc.want, g.normalizeTag())
-		})
-	}
-}
-
 func TestGetAssetAutoDetection(t *testing.T) {
 
 	testData := []struct {
