+++ internal/lib/util_test.go
@@ -29,3 +29,27 @@ func TestReplaceTilde(t *testing.T) {
 		})
 	}
 }
+
+func TestTagNormalization(t *testing.T) {
+	testData := []struct {
+		name string
+		tag  string
+		want string
+	}{
+		{
+			name: "kustomize",
+			tag:  "kustomize/v4.5.4",
+			want: "kustomize-v4.5.4",
+		},
+		{
+			name: "regular",
+			tag:  "v0.21.0",
+			want: "v0.21.0",
+		},
+	}
+	for _, tc := range testData {
+		t.Run(tc.name, func(t *testing.T) {
+			require.Equal(t, tc.want, normalizeTag(tc.tag))
+		})
+	}
+}
